Fix signed/unsigned int confusion when iterating over a GPtrArray.
2004-11-11  Christian Persch  <chpe@cvs.gnome.org>

        * embed/ephy-history.c: (remove_obsolete_pages):
        * src/bookmarks/ephy-bookmarks.c: (clear_favorites):

        Fix signed/unsigned int confusion when iterating over
        a GPtrArray.

+++ embed/ephy-history.c
@@ -231,7 +231,7 @@ remove_obsolete_pages (EphyHistory *eb)
         g_date_set_time (&current_date, time (NULL));
 
 	children = ephy_node_get_children (eb->priv->pages);
-	for (i = children->len -1; i >= 0; i--)
+	for (i = (int) children->len - 1; i >= 0; i--)
 	{
 		EphyNode *kid;
